// src/frontend/src/app/api/reports/swagger.ts
export class GetSuitesTreeResponse implements IGetSuitesTreeResponse {
    suites?: SuitesTreeItem[] | undefined;

    constructor(data?: IGetSuitesTreeResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["suites"])) {
                this.suites = [] as any;
                for (let item of _data["suites"])
                    this.suites!.push(SuitesTreeItem.fromJS(item));
            }
        }
    }

    static fromJS(data: any): GetSuitesTreeResponse {
        data = typeof data === 'object' ? data : {};
        let result = new GetSuitesTreeResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.suites)) {
            data["suites"] = [];
            for (let item of this.suites)
                data["suites"].push(item.toJSON());
        }
        return data;
    }
}

export interface IGetSuitesTreeResponse {
    suites?: SuitesTreeItem[] | undefined;
}

export class SuitesTreeItem implements ISuitesTreeItem {
    id!: number;
    name?: string | undefined;
    parentId?: number | undefined;

    constructor(data?: ISuitesTreeItem) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.parentId = _data["parentId"];
        }
    }

    static fromJS(data: any): SuitesTreeItem {
        data = typeof data === 'object' ? data : {};
        let result = new SuitesTreeItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["parentId"] = this.parentId;
        return data;
    }
}

export interface ISuitesTreeItem {
    id: number;
    name?: string | undefined;
    parentId?: number | undefined;
}
// ...
export class GetSuiteTestPointsResponse implements IGetSuiteTestPointsResponse {
    points?: SuiteTestPointItem[] | undefined;

    constructor(data?: IGetSuiteTestPointsResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["points"])) {
                this.points = [] as any;
                for (let item of _data["points"])
                    this.points!.push(SuiteTestPointItem.fromJS(item));
            }
        }
    }

    static fromJS(data: any): GetSuiteTestPointsResponse {
        data = typeof data === 'object' ? data : {};
        let result = new GetSuiteTestPointsResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.points)) {
            data["points"] = [];
            for (let item of this.points)
                data["points"].push(item.toJSON());
        }
        return data;
    }
}

export interface IGetSuiteTestPointsResponse {
    points?: SuiteTestPointItem[] | undefined;
}

export class SuiteTestPointItem implements ISuiteTestPointItem {
    id!: number;
    testSuiteId!: number;
    testConfigId!: number;
    testCaseId!: number;
    name?: string | undefined;
    order!: number;
    priority!: number;
    outcome!: Outcome;
    configuration?: string | undefined;
    tester?: string | undefined;
    description?: string | undefined;

    constructor(data?: ISuiteTestPointItem) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.testSuiteId = _data["testSuiteId"];
            this.testConfigId = _data["testConfigId"];
            this.testCaseId = _data["testCaseId"];
            this.name = _data["name"];
            this.order = _data["order"];
            this.priority = _data["priority"];
            this.outcome = _data["outcome"];
            this.configuration = _data["configuration"];
            this.tester = _data["tester"];
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): SuiteTestPointItem {
        data = typeof data === 'object' ? data : {};
        let result = new SuiteTestPointItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["testSuiteId"] = this.testSuiteId;
        data["testConfigId"] = this.testConfigId;
        data["testCaseId"] = this.testCaseId;
        data["name"] = this.name;
        data["order"] = this.order;
        data["priority"] = this.priority;
        data["outcome"] = this.outcome;
        data["configuration"] = this.configuration;
        data["tester"] = this.tester;
        data["description"] = this.description;
        return data;
    }
}

export interface ISuiteTestPointItem {
    id: number;
    testSuiteId: number;
    testConfigId: number;
    testCaseId: number;
    name?: string | undefined;
    order: number;
    priority: number;
    outcome: Outcome;
    configuration?: string | undefined;
    tester?: string | undefined;
    description?: string | undefined;
}

export enum Outcome {
    Unknow = 0,
    Planed = 1,
    Passed = 2,
    Failed = 3,
    Blocked = 4,
    Skipped = 5,
    Paused = 6,
}